Split Dao into UserDao & TourDao
