+++ src/main/java/services/TourService.java
@@ -1,31 +1,31 @@
 package services;
 
-import dao.Dao;
+import dao.TourDao;
 import model.Tour;
 
 import java.util.Collection;
 
 public class TourService {
-    private Dao dao;
+    private TourDao dao;
 
     public Collection<Tour> getAllTours(String language) {
-        if (!"RU".equals(language) && "EN".equals(language)) {
+        if ((!"RU".equals(language)) && (!"EN".equals(language))) {
             language = "EN";
         }
         return dao.getAllTours(language);
     }
 
     public void addTour(Tour tour) {
-        if (dao.isExist("tour_details", "title", tour.getTitle())) {
+        if (!dao.isTitleUsed(tour.getTitle())) {
             dao.createTour(tour);
         }
     }
 
     /**
      * This is set via Beans.xml configuration
-     * @param dao
+     * @param dao is dao
      **/
-    public void setDao(Dao dao) {
+    public void setDao(TourDao dao) {
         this.dao = dao;
     }
 }
