+++ src/main/resources/Beans.xml
@@ -13,13 +13,19 @@
         <property name = "password" value = "UDUCVRcm7r"/>
     </bean>
 
-    <!-- Definition for Dao bean -->
-    <bean id = "Dao"
-          class = "dao.DaoImpl">
+    <!-- Definition for UserDao bean -->
+    <bean id = "UserDao"
+          class = "dao.UserDaoImpl">
         <property name = "dataSource" ref = "dataSource" />
     </bean>
 
-    <!-- Definition for RegistrationService bean-->
+    <!-- Definition for TourDao bean -->
+    <bean id = "TourDao"
+          class = "dao.TourDaoImpl">
+        <property name = "dataSource" ref = "dataSource" />
+    </bean>
+
+    <!-- Definition for AuthService bean-->
     <bean id = "authService"
           class="services.AuthService">
         <property name="dao" ref="Dao"/>
