--- src/tests/testResources/Beans.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans
-       http://www.springframework.org/schema/beans/spring-beans.xsd">
-
-    <!-- Initialization for data source -->
-    <bean id="dataSource"
-          class = "org.springframework.jdbc.datasource.DriverManagerDataSource">
-        <property name = "driverClassName" value = "com.mysql.jdbc.Driver"/>
-        <property name = "url" value = "jdbc:mysql://localhost:3306/"/>
-        <property name = "username" value = ""/>
-        <property name = "password" value = ""/>
-    </bean>
-
-    <!-- Definition for Dao bean -->
-    <bean id = "Dao"
-          class = "dao.DaoImpl">
-        <property name = "dataSource" ref = "dataSource" />
-    </bean>
-
-    <!-- Definition for RegistrationService bean-->
-    <bean id = "RegistrationService"
-          class="services.RegistrationService">
-        <property name="dao" ref="Dao"/>
-    </bean>
-
-</beans>
